+++ packages/router/src/bindings/messaging/auctionRequest.ts
@@ -30,7 +30,7 @@ export const auctionRequestBinding = async (
   }
 
   // On every new auction broadcast, route to the new auction handler
-  logger.info("Received auction request", requestContext, methodContext);
+  logger.debug("Received auction request", requestContext, methodContext);
   const { bid, bidSignature, gasFeeInReceivingToken } = await newAuction(data, requestContext);
 
   await messaging.publishAuctionResponse(from, inbox, { bid, bidSignature, gasFeeInReceivingToken });
@@ -40,5 +40,5 @@ export const auctionRequestBinding = async (
     sendingChainId: bid.sendingChainId,
     receivingChainId: bid.receivingChainId,
   });
-  logger.info("Handled auction request", requestContext, methodContext, { bid, gasFeeInReceivingToken });
+  logger.debug("Handled auction request", requestContext, methodContext, { bid, gasFeeInReceivingToken });
 };