+++ packages/router/test/lib/operations/auction.spec.ts
@@ -110,7 +110,7 @@ describe("Auction Operation", () => {
 
       // it should take a gas fee for fulfill transactions if sendingChain is fee chain.
       // amountReceived = amount.sub(fulfillGasFee)
-      const expectedReceiverAmount = "99989650500000000000";
+      const expectedReceiverAmount = "99987400000000000000";
       const bid = await newAuction(auctionPayload, requestContext);
       expect(bid.bid).to.deep.eq({
         user: auctionPayload.user,
@@ -146,7 +146,7 @@ describe("Auction Operation", () => {
 
       // it should take a gas fee for prepare transactions if receivingChain is fee chain.
       // amountReceived = amount.sub(prepareGasFee)
-      const expectedReceiverAmount = "99991342000000000000";
+      const expectedReceiverAmount = "99989500000000000000";
       const bid = await newAuction(auctionPayload, requestContext);
       expect(bid.bid).to.deep.eq({
         user: auctionPayload.user,
@@ -182,7 +182,7 @@ describe("Auction Operation", () => {
 
       // it should take a gas fee for prepare and fulfill transactions if both sendingChain and receivingChain are fee chains.
       // amountReceived = amount.sub(prepareGasFee).sub(fulfillGasFee)
-      const expectedReceiverAmount = "99980992500000000000";
+      const expectedReceiverAmount = "99976900000000000000";
       const bid = await newAuction(auctionPayload, requestContext);
       expect(bid.bid).to.deep.eq({
         user: auctionPayload.user,